rafact: Moving Analyzers to the Analysis module
