+++ py_search_engine/analysis/analyzers/base_analyzer.py
@@ -1,0 +1,9 @@
+class BaseAnalyzer:
+    PIPELINE = []
+
+    def analyze(self, term):
+        terms = [term]
+        for pipeline_step in self.PIPELINE:
+            terms = pipeline_step(terms).process()
+
+        return terms
